Refactored Serialization
