+++ subprojects/semantifyr/src/main/kotlin/utils/IndentationAwareStringWriter.kt
@@ -1,10 +1,10 @@
 /*
- * SPDX-FileCopyrightText: 2023-2024 The Semantifyr Authors
+ * SPDX-FileCopyrightText: 2023-2025 The Semantifyr Authors
  *
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package hu.bme.mit.semantifyr.oxsts.semantifyr.serialization
+package hu.bme.mit.semantifyr.oxsts.semantifyr.utils
 
 class IndentationAwareStringWriter(
     private val indentation: String
@@ -43,6 +43,19 @@ class IndentationAwareStringWriter(
 
     override fun toString() = stringBuilder.toString()
 
+}
+
+inline fun IndentationAwareStringWriter.appendIndent(
+    string: String,
+    openChar: String = " {",
+    closeChar: String = "}",
+    body: IndentationAwareStringWriter.() -> Unit
+) {
+    appendLine(string + openChar)
+    indent {
+        body()
+    }
+    appendLine(closeChar)
 }
 
 inline fun indent(indentation: String = "    ", body: IndentationAwareStringWriter.() -> Unit): String {
